+++ app/ui/settings/browser.js
@@ -1,15 +1,12 @@
 (function () {'use strict';
 
-const electron = require('electron');
-const ipcRenderer = electron.ipcRenderer;
+const { ipcRenderer } = require('electron');
 const handlebars = require('handlebars');
-const uuid4 = require('uuid4');
 
-const clientConfig = require('../../lib/config.js');
-const appState = require('../../lib/state.js');
+const globalConfig = require('../../lib/global-config.js');
+const autoLaunch = require('../../lib/auto-launch.js');
 
 const i18n = require('../../lib/i18n.js');
-const app = electron.remote.app;
 
 handlebars.registerHelper('i18n', function(key) {
   var translation = i18n.__(key);
@@ -22,86 +19,48 @@ $(document).ready(function() {
 
   compileTemplates();
 
-  ipcRenderer.on('update-window', updateWindow);
+  const $navigationItems = $('#settings-navigation li a');
+  const $tabContents = $('.tab-content');
 
-  var $checkForUpdate = $('#check-for-update');
-  var $installUpdate = $('#install-update');
-  var $errorReportErrorMessage = $('#error-report-failure');
-  var $errorReport = $('#settings-report');
+  navigateTo('settings-global');
 
-  $checkForUpdate.bind('click', function() {
-    ipcRenderer.send('check-for-update');
+  $navigationItems.bind('click', function(event) {
+    event.preventDefault();
+
+    navigateTo($(this).attr('href').substr(1));
   });
 
-  $installUpdate.bind('click', function() {
-    ipcRenderer.send('install-update');
-  });
+  function navigateTo(tab) {
+    $navigationItems.parent('li').removeClass('tab-navigation-active');
+    $tabContents.removeClass('tab-active');
 
-
-  $errorReport.bind('click', function() {
-    if($errorReport.is(':visible')) {
-      $errorReport.hide();
-      $errorReportErrorMessage.hide();
-      ipcRenderer.send('settings-send-error-report');
-
-      ipcRenderer.once('settings-send-error-report-result', (event, result) => {
-        $errorReport.show();
-
-        if(result === false) {
-          $errorReportErrorMessage.show();
-        }
-      });
-    }
-  });
-
-  var $usernameValue = $('#value-username');
-  var $logout        = $('#settings-account-logout');
-  var $login         = $('#settings-account-login');
-
-  function updateLoginState() {
-    if(clientConfig.get('loggedin') === true) {
-      $logout.show();
-      $login.hide();
-      $usernameValue.html(clientConfig.get('username'));
-    } else {
-      $logout.hide();
-      $login.show();
-      $usernameValue.html(i18n.__('settings.account.loggedout'));
-    }
+    $navigationItems.addBack().find('[href="#' + tab + '"]').parent('li').addClass('tab-navigation-active');
+    $tabContents.addBack().find('#' + tab).addClass('tab-active');
   }
 
-  function toggleInstallUpdate() {
-    $installUpdate.toggle(appState.get('updateAvailable'));
-    $checkForUpdate.toggle(!appState.get('updateAvailable'));
-  }
+  const $autoLaunchCheck = $('#settings-autolaunch-check');
+  $autoLaunchCheck.attr('checked', autoLaunch.getState());
 
-  function updateWindow() {
-    toggleInstallUpdate();
-    updateLoginState();
-  }
-
-  $logout.bind('click', function() {
-    if($logout.is(':visible')) {
-      $logout.hide();
-      ipcRenderer.send('settings-logout-requested');
-
-      ipcRenderer.once('settings-logout-requested-result', (event, result) => {
-        $logout.show();
-        updateLoginState();
-      });
-    }
-  });
-
-  $login.bind('click', function() {
-    var id = uuid4();
-    ipcRenderer.send('settings-login-requested', id);
-
-    ipcRenderer.once('settings-login-requested-result-'+id, (event, result) => {
-      if(result === true) updateLoginState();
+  $autoLaunchCheck.bind('change', function(event) {
+    this.disabled = true;
+    autoLaunch.setState(this.checked).then((result) => {
+      this.disabled = false;
     });
   });
 
-  updateWindow();
+  const $allowPrereleaseCheck = $('#settings-allowPrerelease-check');
+  $allowPrereleaseCheck.attr('checked', globalConfig.get('allowPrerelease'));
+
+  $allowPrereleaseCheck.bind('change', function(event) {
+    globalConfig.set('allowPrerelease', this.checked);
+  });
+
+
+
+  $('#settings-close').bind('click', function(event) {
+    event.preventDefault();
+    ipcRenderer.send('settings-close');
+  });
 });
 
 function compileTemplates() {
@@ -109,10 +68,7 @@ function compileTemplates() {
   var $placeholderContent = $('#contentWrapper');
   var templateContent = handlebars.compile(templateContentHtml);
 
-  var context = {
-    version: app.getVersion(),
-    balloonDir: clientConfig.get('balloonDir')
-  };
+  var context = {};
 
   $placeholderContent.html(templateContent(context));
 }