+++ app/ui/tray/browser.js
@@ -84,6 +84,12 @@ function buildMenu() {
     }}))
   }
 
+  label = i18n.__('tray.menu.settings');
+  menu.append(new MenuItem({label: label, click: function(){
+    ipcRenderer.send('settings-open');
+    ipcRenderer.send('tray-hide');
+  }}))
+
   label = i18n.__('tray.menu.feedback');
   menu.append(new MenuItem({label: label, click: function(){
     ipcRenderer.send('feedback-open');