+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * CORE: [CHANGE] introduced global app state, moved updateAvailable and onLineState to global app state #26
 * CORE: [CHANGE] introduced new global flag `allowPrerelease` if set to true auto updater will install pre releases.
+* CORE: [CHANGE] introduced settings dialog #56
 
 ## 0.0.41
 **Maintainer**: balloon-team <opensource@gyselroth.net>\