+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 **Date**: Mon Feb 26 16:05:52 CET 2018
 
 * CORE: [CHANGE] introduced global app state, moved updateAvailable and onLineState to global app state #26
-* CORE: [CHANGE] introduced new global flag `allowPrerelease` if set to true auto updater will install pre releases.
+* CORE: [CHANGE] introduced new global flag `allowPrerelease` if set to true auto updater will install pre releases. #55
 
 
 ## 0.0.41