+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.0.43-beta.1
+
 ## 0.0.42
 **Maintainer**: balloon-team <opensource@gyselroth.net>\
 **Date**: Mon Feb 26 16:05:52 CET 2018